# Optimize engine construction