--- a/Jint.Benchmark/EngineConstructionBenchmark.cs
+++ b/Jint.Benchmark/EngineConstructionBenchmark.cs
@@ -1,24 +1,36 @@
 ﻿using BenchmarkDotNet.Attributes;
 using Esprima;
 using Esprima.Ast;
+using Jint.Native;
 
 namespace Jint.Benchmark;
 
 [MemoryDiagnoser]
 public class EngineConstructionBenchmark
 {
-    private readonly Script _program;
+    private Script _program;
+    private Script _simple;
 
-    public EngineConstructionBenchmark()
+    [GlobalSetup]
+    public void GlobalSetup()
     {
         var parser = new JavaScriptParser();
-        _program = parser.ParseScript("return [].length + ''.length");
+        _program = parser.ParseScript("([].length + ''.length)");
+        _simple = parser.ParseScript("1");
+        new Engine().Evaluate(_program);
     }
 
     [Benchmark]
-    public double BuildEngine()
+    public Engine BuildEngine()
     {
         var engine = new Engine();
-        return engine.Evaluate(_program).AsNumber();
+        return engine;
+    }
+
+    [Benchmark]
+    public JsValue EvaluateSimple()
+    {
+        var engine = new Engine();
+        return engine.Evaluate(_simple);
     }
 }