… Jint/Native/Global/GlobalObject.cs