* reuse default options instance if no configure defined
* resolve default time zone and culture only once for options
* pre-calculate global keys (hash codes), use faster path for dictionary init

## Jint.Benchmark.EngineConstructionBenchmark

| **Diff**|Method|Mean|Error|Allocated|
|------- |-------|-------:|-------|-------:|
| Old |BuildEngine|2.033 μs|0.0198 μs|11.33 KB|
| **New** |	| **1.310 μs (-36%)** | **0.0105 μs** | **10.62 KB (-6%)** |
| Old |EvaluateSimple|2.382 μs|0.0287 μs|12.27 KB|
| **New** |	| **1.632 μs (-31%)** | **0.0154 μs** | **11.55 KB (-6%)** |